Excel Files/requirements.txt

streamlit
google-genai
openpyxl
pandas
numpy
python-dotenv
plotly